backend/requirements.txt
fastapi
uvicorn[standard]
paho-mqtt>=2.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-multipart==0.0.6
python-dotenv==1.0.1